requirements.txt
streamlit
langchain
langchain-groq
langchain-openai
langchain-huggingface
sentence-transformers
requests
beautifulsoup4
certifi
python-dotenv
playwright